# spec/openapi.yml
...
openapi: 3.0.2
info:
  title: Marquez
  version: 0.51.2-SNAPSHOT
  description: Marquez is an open source **metadata service** for the **collection**, **aggregation**,
    and **visualization** of a data ecosystem's metadata.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:5000/api/v1
    description: Local API server
paths:
...
  /events/lineage:
    get:
      operationId: getLineageEvents
      parameters:
        - $ref: '#/components/parameters/sortDirection'
        - $ref: '#/components/parameters/before'
        - $ref: '#/components/parameters/after'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all received OpenLineage events.
      description: Returns a list of OpenLineage events, sorted in direction of passed sort parameter. By default it is desc.
      tags:
        - Events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageEventList'
  /sources/{source}:
    parameters:
      - $ref: '#/components/parameters/source'
    put:
      operationId: putSource
      summary: Create a source
      description: Creates a new source object. A source is the physical location of a dataset such as
        a table in PostgreSQL, or topic in Kafka. A source enables the grouping of physical datasets
        to their physical source.
      deprecated: true
      tags:
        - Sources
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedSource'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
    get:
      operationId: getSource
      summary: Retrieve a source
      description: Returns a source.
      tags:
        - Sources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
  /sources:
    get:
      operationId: getSources
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all sources
      description: Returns a list of sources.
      tags:
        - Sources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceList'
...
  /lineage:
    post:
      operationId: recordLineage
      summary: Record a single lineage event
      description: >-
        Receive and commit a single lineage event to a durable database-backed intake queue using the
        [OpenLineage](https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.json) standard.
        Derived lineage projections are processed asynchronously and may not be visible immediately.
        There is no application-level event-count cap; database capacity applies. Acceptance confirms
        only that the event was committed to the intake queue. The queue retries database projection
        failures and can ultimately dead-letter an event after repeated failures. Search indexing and
        run-transition listener delivery happen after the queue item is acknowledged and are best
        effort: they have no durable delivery guarantee, may be missing, concurrent, or out of order,
        and are not retried by the intake queue. Any 5xx response during queue admission is
        commit-indeterminate: the queue transaction may already have committed even if Marquez did
        not observe the commit result. Clients that retry must tolerate duplicate events.
      tags:
        - Lineage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenLineageEventRequest'
      responses:
        '201':
          description: Empty response confirming only that the lineage event was committed to the durable intake queue
        '400':
          description: Malformed lineage event JSON, invalid input dataset identifier, or a text value containing U+0000
        '422':
          description: Lineage event failed request validation
        '500':
          description: >-
            Queue admission failed or its commit outcome is indeterminate; the lineage event may
            already have been committed, and a retry can submit a duplicate
    get:
      operationId: getLineage
      parameters:
        - $ref: '#/components/parameters/nodeId'
        - $ref: '#/components/parameters/depth'
      tags:
        - Lineage
      summary: Get a lineage graph
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageGraph'
...
components:
  parameters:
...
    source:
      name: source
      in: path
      description: The name of the source.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-source
...
    limit:
      name: limit
      in: query
      description: The number of results to return from offset.
      required: false
      schema:
        type: integer
        example: 25
        default: 100
    offset:
      name: offset
      in: query
      description: The initial position from which to return results.
      required: false
      schema:
        type: integer
        default: 0
    nodeId:
      name: nodeId
      in: query
      schema:
        type: string
        example: dataset:food_delivery:public.delivery_7_days
      description: The ID of the node. A node can either be a dataset node, a dataset field node or a job node.
        The format of nodeId for dataset is `dataset:<namespace_of_dataset>:<name_of_the_dataset>`, for dataset field
        is `datasetField:<namespace_of_dataset>:<name_of_the_dataset>:<name_of_field>`, and for job is
        `job:<namespace_of_the_job>:<name_of_the_job>`.
      required: true
    depth:
      name: depth
      in: query
      schema:
        type: integer
        default: 20
      description: Depth of lineage graph to create.
      required: false
...
    sortDirection:
      name: sortDirection
      in: query
      schema:
        type: string
        example: name
        description: Sorts the results of your query by indicated direction `asc` or `desc`.
      required: false
    before:
      name: before
      in: query
      description: Returns events before passed date.
      required: false
      schema:
        type: string
        format: date-time
        example: 2022-09-15T07:47:19Z
    after:
      name: after
      in: query
      description: Returns events after passed date.
      required: false
      schema:
        type: string
        format: date-time
        example: 2022-09-15T07:47:19Z
...
  schemas:
...
    LineageEventList:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/LineageEvent'
        totalCount:
          type: number
          description: The total number of events returned matching our conditions.
    CreatedSource:
      type: object
      properties:
        type:
          description: The type of the source.
          type: string
        connectionUrl:
          description: The URL to the location of the source.
          type: string
          format: URL
        description:
          description:  The description of the source.
          type: string
      required:
        - type
        - connectionUrl
      example:
        type: POSTGRESQL
        connectionUrl: jdbc:postgresql://db.example.com/mydb
        description: My first source!
    Source:
      type: object
      properties:
        type:
          description: The type of the source.
          type: string
        name:
          description: The name of the source.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the source was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the source was updated.
          type: string
          format: date-time
        connectionUrl:
          description: The URL to the location of the source.
          type: string
          format: URL
        description:
          description:  The description of the dataset.
          type: string
      example:
        type: POSTGRESQL
        name: my-source
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        connectionUrl: jdbc:postgresql://db.example.com/mydb
        description: My first source!
    SourceList:
      type: object
      properties:
        sources:
          items:
            $ref: '#/components/schemas/Source'
...
    DatasetId:
      type: object
      description: The ID of the dataset.
      properties:
        namespace:
          type: string
          description: The namespace of the dataset.
        name:
          type: string
          description: The name of the dataset.
    Dataset:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/DatasetId'
        type:
          description: The type of the dataset.
          type: string
        name:
          description: The **logical** name of the dataset.
          type: string
        physicalName:
          description: The **physical** name of the dataset.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the dataset was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the dataset was updated.
          type: string
          format: date-time
        namespace:
          type: string
          description: The namespace of the dataset.
        sourceName:
          description: The name of the source associated with the dataset.
          type: string
        fields:
          description: The fields of the dataset.
          type: array
          items:
            type: object
            properties:
              name:
                description: The name of the field.
                type: string
              type:
                description: The data type of the field.
                type: string
              tags:
                description: List of tags.
                type: array
                items:
                  type: string
              description:
                description: The description of the field.
                type: string
            required:
              - name
              - type
        tags:
          description: List of tags.
          type: array
          items:
            type: string
        lastModifiedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the dataset was last modified by a successful run.
          type: string
          format: date-time
        lastLifecycleState:
          description: The last lifecycle state of the dataset.
          type: string
        description:
          description: The description of the dataset.
          type: string
        facets:
          $ref: '#/components/schemas/DatasetFacets'
        currentVersion:
          description: The current version of the dataset.
          type: string
          format: uuid
        deleted:
          description: The deleted state of the dataset.
          type: boolean
      example:
        id:
          namespace: my-namespace
          name: my-dataset
        type: DB_TABLE
        name: my-dataset
        physicalName: public.mytable
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        namespace: my-namespace
        sourceName: my-source
        fields: [
          {name': a, type: INTEGER, tags: []},
          {name': b, type: TIMESTAMP, tags: []},
          {name': c, type: INTEGER, tags: []},
          {name': d, type: INTEGER, tags: []}
        ]
        tags: []
        lastModifiedAt: null
        description: "My first dataset!"
        facets: {}
        currentVersion: "b1d626a2-6d3a-475e-9ecf-943176d4a8c6"
...
    JobId:
      description: The ID of the job.
      type: object
      properties:
        namespace:
          type: string
          description: The namespace of the job.
        name:
          type: string
          description: The name of the job.
    Job:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/JobId'
        type:
          description: The type of the job.
          type: string
        name:
          description: The name of the job.
          type: string
        simpleName:
          description: The shortened name of the job.
          type: string
        parentJobName:
          description: The name of the parent job.
          type: string
        parentJobUuid:
          description: The UUID of the parent job.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the job was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the job was updated.
          type: string
          format: date-time
        namespace:
          description: The namespace of the job.
          type: string
        inputs:
          description: The set of **input** datasets.
          type: array
          items:
            $ref: '#/components/schemas/DatasetId'
          uniqueItems: true
        outputs:
          description: The set of **output** datasets.
          type: array
          items:
            $ref: '#/components/schemas/DatasetId'
          uniqueItems: true
        location:
          description: The URL of the job source code or artifact.
          type: string
          format: URL
        context:
          description: A key/value pair that must be of type `string`. A context can be used for getting additional details about the job.
          type: object
          deprecated: true
        description:
          description: The description of the job.
          type: string
        latestRun:
          $ref: '#/components/schemas/Run'
        latestRuns:
          type: array
          items:
            $ref: '#/components/schemas/Run'
        facets:
          $ref: '#/components/schemas/JobFacets'
        currentVersion:
          description: The current version of the job.
          type: string
          format: uuid
      example:
        id:
          namespace: my-namespace
          name: my-job
        type: BATCH
        name: my-job
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        namespace: my-namespace
        inputs: [{namespace: my-namespace, name: my-dataset}]
        outputs: []
        location: https://github.com/my-jobs/blob/124f6089ad4c5fcbb1d7b33cbb5d3a9521c5d32c
        context: {'SQL': "SELECT * FROM mytable;"}
        description: My first job!
        latestRun: null
        latestRuns: []
        facets: {}
        currentVersion: "b1d626a2-6d3a-475e-9ecf-943176d4a8c6"
...
    JobVersionId:
      type: object
      properties:
        namespace:
          type: string
          description: The namespace of the job.
        name:
          type: string
          description: The name of the job.
        version:
          type: string
          format: uuid
          description: The version of the job.
...
    IncompleteRun:
      type: object
      properties:
        id:
          description: The ID of the run.
          type: string
          format: uuid
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the run was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the run was updated.
          type: string
          format: date-time
        nominalStartTime:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the nominal start time of the run.
          type: string
          format: date-time
        nominalEndTime:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the nominal end time of the run.
          type: string
          format: date-time
        state:
          description: The current state of the run.
          type: string
          enum: [NEW, RUNNING, COMPLETED, FAILED, ABORTED]
        startedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the start time of the run.
          type: string
          format: date-time
        endedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the end time of the run.
          type: string
          format: date-time
        durationMs:
          type: integer
          description: The total duration of the run. Measured in milliseconds.
        args:
          description: The arguments of the run.
          type: object
        facets:
          $ref: '#/components/schemas/RunFacets'
      example:
        id: 870492da-ecfb-4be0-91b9-9a89ddd3db90
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        nominalStartTime: null
        nominalEndTime: null
        state: RUNNING
        startedAt: 2019-05-09T15:17:32.690346
        endedAt: null
        durationMs: null
        args: {'email': 'me@example.com', 'emailOnFailure': 'false', 'emailOnRetry': 'true', 'retries': '1'}
        facets: {}
    Run:
      type: object
      anyOf:
        - $ref: '#/components/schemas/IncompleteRun'
        - type: object
          properties:
            jobVersion:
              $ref: '#/components/schemas/JobVersionId'
            inputDatasetVersions:
              type: array
              items:
                $ref: '#/components/schemas/InputDatasetVersion'
            outputDatasetVersions:
              type: array
              items:
                $ref: '#/components/schemas/OutputDatasetVersion'
            context:
              description: A key/value pair that must be of type `string`. A context can be used for getting additional details about the job.
              type: object
              deprecated: true
            facets:
              $ref: '#/components/schemas/RunFacets'
      example:
        id: 870492da-ecfb-4be0-91b9-9a89ddd3db90
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        nominalStartTime: null
        nominalEndTime: null
        state: COMPLETED
        startedAt: 2019-05-09T15:17:32.690346
        endedAt: 2019-05-09T20:05:46.815920Z
        durationMs: 4250894125
        args: {'email': 'me@example.com', 'emailOnFailure': 'false', 'emailOnRetry': 'true', 'retries': '1'}
        context: {'SQL': "SELECT * FROM mytable;"}
        facets: {}
...
    DatasetFacets:
      description: The facets of the dataset. For a full list of standard dataset facets, see [OpenLineage](https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.md#standard-facets).
      type: object
      additionalProperties:
        $ref: '#/components/schemas/CustomFacet'
    JobFacets:
      description: The facets of the job. For a full list of standard job facets, see [OpenLineage](https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.md#standard-facets).
      type: object
      additionalProperties:
        $ref: '#/components/schemas/CustomFacet'
    RunFacets:
      description: The facets of the run. For a full list of standard run facets, see [OpenLineage](https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.md#standard-facets).
      type: object
      additionalProperties:
        $ref: '#/components/schemas/CustomFacet'
    CustomFacet:
      description: A custom facet enables the extension of _dataset_, _job_, and _run_ metadata. A custom facet **must** also have a schema,
        where a version of the schema is identifiable via a URL. A field within the schema **must** not start with an underscore (`_`).
      allOf:
        - $ref: '#/components/schemas/BaseFacet'
        - type: object
          additionalProperties: true
...
    LineageGraph:
      type: object
      properties:
        graph:
          type: array
          items:
            $ref: '#/components/schemas/GraphNode'
    GraphNode:
      type: object
      properties:
        id:
          type: string
          description: The ID of the node.
        type:
          $ref: '#/components/schemas/NodeType'
        data:
          type: object
          description: The data of the node.
          discriminator:
            propertyName: type
            mapping:
              'DATASET': '#/components/schemas/Dataset'
              'JOB': '#/components/schemas/Job'
          oneOf:
            - $ref: '#/components/schemas/Dataset'
            - $ref: '#/components/schemas/Job'
        inEdges:
          type: array
          description: The in edges of the node.
          items:
            $ref: '#/components/schemas/GraphEdge'
        outEdges:
          type: array
          description: The out edges of the node.
          items:
            $ref: '#/components/schemas/GraphEdge'
    NodeType:
      type: string
      enum: [ JOB, DATASET ]
      description: The type of the diff node.
    GraphEdge:
      type: object
      properties:
        origin:
          description: The ID of the origin node.
          type: string
        destination:
          description: The ID of the destination node.
          type: string
...
    OpenLineageEventRequest:
      description: >-
        A supported OpenLineage RunEvent, DatasetEvent, or JobEvent. The variants are selected by
        their documented shape and schema URL; no discriminator is used because a RunEvent may omit
        schemaURL for backward compatibility.
      oneOf:
        - $ref: '#/components/schemas/OpenLineageRunEventRequest'
        - $ref: '#/components/schemas/OpenLineageDatasetEventRequest'
        - $ref: '#/components/schemas/OpenLineageJobEventRequest'
    OpenLineageRunEventRequest:
      type: object
      additionalProperties: true
      required: [eventTime, run, job, producer]
      properties:
        eventType:
          type: string
        eventTime:
          type: string
          format: date-time
        run:
          $ref: '#/components/schemas/OpenLineageRunRequest'
        job:
          $ref: '#/components/schemas/OpenLineageJobRequest'
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/OpenLineageDatasetRequest'
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/OpenLineageDatasetRequest'
        producer:
          type: string
        schemaURL:
          type: string
          format: uri
          nullable: true
          pattern: '/RunEvent$'
          description: May be omitted or null; the server then uses the RunEvent model.
      example:
        eventType: COMPLETE
        eventTime: 2019-05-09T19:49:24.201361Z
        run: {runId: d46e465b-d358-4d32-83d4-df660ff614dd}
        job: {namespace: my-namespace, name: my-job}
        outputs: [{namespace: my-namespace, name: my-output}]
        producer: https://example.com/openlineage-producer
    OpenLineageDatasetEventRequest:
      type: object
      additionalProperties: true
      required: [eventTime, dataset, producer, schemaURL]
      properties:
        eventTime:
          type: string
          format: date-time
        dataset:
          $ref: '#/components/schemas/OpenLineageDatasetRequest'
        producer:
          type: string
        schemaURL:
          type: string
          format: uri
          pattern: '/DatasetEvent$'
      example:
        eventTime: 2019-05-09T19:49:24.201361Z
        dataset: {namespace: my-namespace, name: my-dataset}
        producer: https://example.com/openlineage-producer
        schemaURL: https://openlineage.io/spec/2-0-0/OpenLineage.json#/definitions/DatasetEvent
    OpenLineageJobEventRequest:
      type: object
      additionalProperties: true
      required: [eventTime, job, producer, schemaURL]
      properties:
        eventTime:
          type: string
          format: date-time
        job:
          $ref: '#/components/schemas/OpenLineageJobRequest'
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/OpenLineageDatasetRequest'
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/OpenLineageDatasetRequest'
        producer:
          type: string
        schemaURL:
          type: string
          format: uri
          pattern: '/JobEvent$'
      example:
        eventTime: 2019-05-09T19:49:24.201361Z
        job: {namespace: my-namespace, name: my-job}
        producer: https://example.com/openlineage-producer
        schemaURL: https://openlineage.io/spec/2-0-0/OpenLineage.json#/definitions/JobEvent
    OpenLineageRunRequest:
      type: object
      additionalProperties: true
      required: [runId]
      properties:
        runId:
          $ref: '#/components/schemas/OpenLineageIdentityRequest'
        facets:
          $ref: '#/components/schemas/OpenLineageRunFacetsRequest'
    OpenLineageRunFacetsRequest:
      type: object
      additionalProperties: true
      properties:
        parent:
          $ref: '#/components/schemas/OpenLineageParentRunFacetRequest'
        parentRun:
          $ref: '#/components/schemas/OpenLineageParentRunFacetRequest'
    OpenLineageParentRunFacetRequest:
      type: object
      additionalProperties: true
      required: [_producer, _schemaURL, run, job]
      properties:
        _producer:
          type: string
          format: uri
        _schemaURL:
          type: string
          format: uri
        run:
          $ref: '#/components/schemas/OpenLineageRunLinkRequest'
        job:
          $ref: '#/components/schemas/OpenLineageJobRequest'
    OpenLineageRunLinkRequest:
      type: object
      additionalProperties: true
      required: [runId]
      properties:
        runId:
          $ref: '#/components/schemas/OpenLineageIdentityRequest'
    OpenLineageJobRequest:
      type: object
      additionalProperties: true
      required: [namespace, name]
      properties:
        namespace:
          $ref: '#/components/schemas/OpenLineageIdentityRequest'
        name:
          $ref: '#/components/schemas/OpenLineageIdentityRequest'
        facets:
          type: object
          additionalProperties: true
    OpenLineageDatasetRequest:
      type: object
      additionalProperties: true
      required: [namespace, name]
      properties:
        namespace:
          $ref: '#/components/schemas/OpenLineageIdentityRequest'
        name:
          $ref: '#/components/schemas/OpenLineageIdentityRequest'
        facets:
          type: object
          additionalProperties: true
        inputFacets:
          type: object
          additionalProperties: true
        outputFacets:
          type: object
          additionalProperties: true
    OpenLineageIdentityRequest:
      type: string
      minLength: 1
      description: >-
        An OpenLineage identity component. Runtime validation also rejects values made entirely of
        Java or Unicode separator whitespace, which OpenAPI 3.0 cannot express portably.
    LineageEvent:
      example:
        eventType: COMPLETE
        eventTime: 2019-05-09T19:49:24.201361Z
        run: {runId: d46e465b-d358-4d32-83d4-df660ff614dd}
        job: {namespace: my-namespace,name: my-job}
        outputs: [
          {
            namespace: my-namespace,
            name: my-output,
            facets: {
              schema: {
                _producer: "https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/client",
                _schemaURL: "https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/spec/OpenLineage.json#/definitions/SchemaDatasetFacet",
                fields: [
                  { name: a, type: INTEGER },
                  { name: b, type: TIMESTAMP },
                  { name: c, type: INTEGER },
                  { name: d, type: INTEGER }
                ]
              }
            }
          }
        ]
        producer: "https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/client"
...
    InputDatasetVersion:
      type: object
      properties:
        datasetVersionId:
          $ref: '#/components/schemas/DatasetVersionId'
        facets:
          type: object
          additionalProperties:
            type: string
          description: Dataset facets in run context, like `inputFacets`.
      required:
        - datasetVersionId
        - facets
    OutputDatasetVersion:
      type: object
      properties:
        datasetVersionId:
          $ref: '#/components/schemas/DatasetVersionId'
        facets:
          type: object
          additionalProperties:
            type: string
          description: Dataset facets in run context, like `outputFacets`.
      required:
        - datasetVersionId
        - facets